llm/requirements.txt
aiohttp==3.10.10
arxiv==2.1.3
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
blobfile==3.0.0
dataclasses-json==0.6.7
fastapi==0.115.2
transformers==4.45.2
huggingface-hub==0.26.0
fire==0.7.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
jaraco.collections==5.1.0
langchain-text-splitters==0.3.0
llama-models==0.0.45
motor==3.6.0
pip-chill==1.0.3
platformdirs==4.2.2
playwright==1.47.0
prompt-toolkit==3.0.48
pydantic-settings==2.6.0
rich==13.9.3
sentence-transformers==3.2.0
sqlalchemy==2.0.36
tavily-python==0.5.0
pymongo==4.9.2
tomli==2.0.1
uvicorn==0.32.0
